- added basic loading ui test
