+++ Xamarin.Forms.Controls.Issues/Xamarin.Forms.Controls.Issues.Shared/VisualControlsPage.xaml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <ContentPage xmlns="http://xamarin.com/schemas/2014/forms"
             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-            x:Class="Xamarin.Forms.Controls.VisualGallery"
+            x:Class="Xamarin.Forms.Controls.Issues.VisualControlsPage"
             Visual="Material">
 
     <ContentPage.Resources>
