+++ Xamarin.Forms.Controls/ControlGalleryPages/VisualGallery.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+
+using Xamarin.Forms;
+using Xamarin.Forms.Controls.Issues;
+
+namespace Xamarin.Forms.Controls
+{
+	public class VisualGallery : VisualControlsPage
+	{
+		public VisualGallery()
+		{
+		}
+	}
+}
