+++ Xamarin.Forms.Core.UITests.Shared/UITestCategories.cs
@@ -45,5 +45,6 @@
 
 		public const string ManualReview = "ManualReview";
 		public const string Performance = "Performance";
+		public const string Visual = "Visual";
 	}
 }
